requirements.txt
Flask==3.0.2
Typer==0.12.3
SQLAlchemy==2.0.30
bcrypt==4.1.2
uvicorn==0.29.0
python-dotenv==1.0.1
pymysql
passlib